Remove symbol from GetOrderOcoOptions
GetOrderOptions has a `symbol` property that is converted as queryParam when requesting Binance API.
The API documentation states that `symbol` is not part of the valid parameters.
Since this property is mandatory on the type definition and Binance rejects the request with an error, this makes it impossible to use this method on applications.

Therefore I'm removing this property from the type definition. 
Tested locally.